refactored logic up to deal cards
