+++ backend/utils/setDealCardsTimeStamp.js
@@ -1,0 +1,22 @@
+const { rooms } = require('../global');
+
+
+function setDealCardsTimeStamp(io, tableId) {
+  if (!rooms[tableId]) return;
+  if (rooms[tableId].dealCardsTimeStamp) return;
+
+  let room = tableId;
+  // Set countdown end time
+  const countdownDuration = 1000; // 5 seconds
+  const endTime = Math.ceil(Date.now() + countdownDuration);
+  rooms[tableId].dealCardsTimeStamp = endTime;
+
+  let countdownObj = {
+    dealCardsTimeStamp: endTime,
+    tableId,
+  };
+
+  io.in(room).emit('countdown_update', countdownObj);
+}
+
+module.exports = setDealCardsTimeStamp;
